# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Learning Management System API
  description: LMS API for managing courses, modules, lessons, enrollments and progress tracking
  version: 1.0.0
  contact:
    name: LMS Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Development
  - url: /api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CourseCreateRequest:
      type: object
      required:
        - courseTitle
        - instructorId
      properties:
        courseTitle:
          type: string
          maxLength: 100
          minLength: 1
          example: Java Spring Boot
        courseDescription:
          type: string
          maxLength: 200
          minLength: 1
          example: Complete Spring Boot course
        courseTags:
          type: array
          items:
            type: string
          default: []
          example:
            - java
            - spring
            - backend
        instructorId:
          type: integer
          format: int64
          example: 101
        canEnrollment:
          type: boolean
          default: false
          example: true
    CourseUpdateRequest:
      type: object
      properties:
        courseTitle:
          type: string
          maxLength: 100
          minLength: 1
        courseDescription:
          type: string
          maxLength: 200
          minLength: 1
        courseTags:
          type: array
          items:
            type: string
          default: []
        courseStatus:
          $ref: './openapi/common.yaml#/components/schemas/CourseStatus'
        canEnrollment:
          type: boolean
    CourseResponse:
      type: object
      properties:
        courseId:
          type: integer
          format: int64
        courseTitle:
          type: string
        courseDescription:
          type: string
        courseTags:
          type: array
          items:
            type: string
          default: []
        courseStatus:
          $ref: './openapi/common.yaml#/components/schemas/CourseStatus'
        instructorId:
          type: integer
          format: int64
        canEnrollment:
          type: boolean
        createdAt:
          type: string
          format: date-time
        modules:
          type: array
          items:
            $ref: './openapi/common.yaml#/components/schemas/ModuleSummary'
          default: []
    CourseListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: './openapi/common.yaml#/components/schemas/CourseSummary'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    LessonCreateRequest:
      type: object
      required:
        - moduleId
        - contentType
        - contentLink
        - sequence
      properties:
        moduleId:
          type: integer
          format: int64
          example: 10
        contentType:
          $ref: './openapi/common.yaml#/components/schemas/ContentType'
          example: MP4
        contentLink:
          type: string
          format: uri
          example: https://cdn.example.com/videos/introduction.mp4
        sequence:
          type: integer
          minimum: 1
          example: 1
    LessonUpdateRequest:
      type: object
      properties:
        contentType:
          $ref: './openapi/common.yaml#/components/schemas/ContentType'
        contentLink:
          type: string
          format: uri
        sequence:
          type: integer
          minimum: 1
        isActive:
          type: boolean
    LessonResponse:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        moduleId:
          type: integer
          format: int64
        contentType:
          $ref: './openapi/common.yaml#/components/schemas/ContentType'
        contentLink:
          type: string
          format: uri
        sequence:
          type: integer
        userId:
          type: integer
          format: int64
          nullable: true
        lessonStatus:
          $ref: './openapi/common.yaml#/components/schemas/LessonStatus'
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    LessonListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: './openapi/common.yaml#/components/schemas/LessonSummary'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
    EnrollmentCreateRequest:
      type: object
      required:
        - userId
        - courseId
      properties:
        userId:
          type: integer
          format: int64
          example: 25
        courseId:
          type: integer
          format: int64
          example: 10
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        courseTitle:
          type: string
        courseCompletionStatus:
          $ref: './openapi/common.yaml#/components/schemas/CourseCompletionStatus'
        enrolledAt:
          type: string
          format: date-time
    EnrollmentListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/EnrollmentResponse'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
  /courses:
    post:
      tags:
        - Courses
      summary: Create a new course
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                type: object
                title: CourseCreateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/CourseResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: './openapi/common.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    get:
      tags:
        - Courses
      summary: List courses
      operationId: listCourses
      parameters:
        - name: courseStatus
          in: query
          description: Filter courses by lifecycle status. If omitted, all statuses are returned.
          schema:
            $ref: './openapi/common.yaml#/components/schemas/CourseStatus'
        - name: pageNo
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: Courses fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /courses/{courseId}:
    get:
      tags:
        - Courses
      summary: Get course by ID
      operationId: getCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: includeModules
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Course fetched successfully
          content:
            application/json:
              schema:
                type: object
                title: CourseGetResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/CourseResponse'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Not found

    put:
      tags:
        - Courses
      summary: Update course
      operationId: updateCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                type: object
                title: CourseUpdateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/CourseResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - Courses
      summary: Delete course
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Course deleted successfully
          content:
            application/json:
              schema:
                type: object
                title: CourseDeleteResponse
                properties:
                  message:
                    type: string
                  data:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict
...
  /lessons:
    post:
      tags:
        - Lessons
      summary: Create a new lesson
      operationId: createLesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreateRequest'
      responses:
        '201':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                type: object
                title: LessonCreateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LessonResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict

    get:
      tags:
        - Lessons
      summary: List lessons
      operationId: listLessons
      parameters:
        - name: moduleId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
        - name: active
          in: query
          schema:
            type: boolean
            default: true
        - name: pageNo
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: sequence
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        '200':
          description: Lessons fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonListResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /lessons/{lessonId}:
    get:
      tags:
        - Lessons
      summary: Get lesson by ID
      operationId: getLesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lesson fetched successfully
          content:
            application/json:
              schema:
                type: object
                title: LessonGetResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LessonResponse'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Not found

    put:
      tags:
        - Lessons
      summary: Update lesson
      operationId: updateLesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdateRequest'
      responses:
        '200':
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                type: object
                title: LessonUpdateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LessonResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict

    delete:
      tags:
        - Lessons
      summary: Delete lesson
      operationId: deleteLesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lesson deleted successfully
          content:
            application/json:
              schema:
                type: object
                title: LessonDeleteResponse
                properties:
                  message:
                    type: string
                  data:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict
  /enrollments:
    post:
      tags:
        - Enrollments
      summary: Create enrollment
      operationId: createEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentCreateRequest'
      responses:
        '201':
          description: Enrollment created successfully
          content:
            application/json:
              schema:
                type: object
                title: EnrollmentCreateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EnrollmentResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict

    get:
      tags:
        - Enrollments
      summary: List enrollments
      operationId: listEnrollments
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: courseCompletionStatus
          in: query
          schema:
            type: string
            enum:
              - COMPLETE
              - INCOMPLETE
        - name: pageNo
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: enrolledAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: Enrollments fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentListResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /enrollments/{enrollmentId}:
    get:
      tags:
        - Enrollments
      summary: Get enrollment by ID
      operationId: getEnrollment
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Enrollment fetched successfully
          content:
            application/json:
              schema:
                type: object
                title: EnrollmentGetResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EnrollmentResponse'
                  timestamp:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - Enrollments
      summary: Cancel enrollment
      operationId: cancelEnrollment
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Enrollment cancelled successfully
          content:
            application/json:
              schema:
                type: object
                title: EnrollmentDeleteResponse
                properties:
                  message:
                    type: string
                  data:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: Not found
...